# Conflicts:
#	src/main/java/model/CanadaGame.java
#	src/main/java/model/CanadaPhysics.java
#	src/main/java/model/GameObjectFactory.java
#	src/main/java/model/world/World.java

+++ src/main/java/engine/IGame.java
@@ -12,7 +12,7 @@ public interface IGame {
 	 * methode qui contient l'evolution du jeu
 	 * @param dt
 	 */
-	public void evolve(float dt);
+	public void evolve(double dt);
 
 	/**
 	 * @return true si et seulement si le jeu est fini
